refactor(annil): move more code from main to modules
