+++ annil/Cargo.toml
@@ -18,7 +18,7 @@ anni-flac = { path = "../anni-flac", version = "0.2.2", features = ["async"] }
 anni-repo = { path = "../anni-repo", version = "0.1.1", features = [
     "git",
     "db-write",
-] }
+], optional = true }
 anni-provider = { path = "../anni-provider", version = "0.1.1" }
 
 serde = { version = "1.0", features = ["derive"] }
@@ -31,4 +31,4 @@ base64 = "0.20.0"
 
 [features]
 default = ["metadata"]
-metadata = []
+metadata = ["anni-repo"]
