--- annil/src/error.rs
@@ -1,21 +1,0 @@
-use axum::http::StatusCode;
-use axum::response::{IntoResponse, Response};
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum AnnilError {
-    #[error("unauthorized")]
-    Unauthorized,
-    #[error("not found")]
-    NotFound,
-}
-
-impl IntoResponse for AnnilError {
-    fn into_response(self) -> Response {
-        match self {
-            AnnilError::Unauthorized => StatusCode::FORBIDDEN,
-            AnnilError::NotFound => StatusCode::NOT_FOUND,
-        }
-        .into_response()
-    }
-}
