+++ annil/src/main.rs
@@ -1,13 +1,12 @@
-use annil::config::{Config, MetadataConfig, ProviderItem};
-use annil::provider::AnnilProvider;
-use annil::utils::compute_etag;
+use config::{Config, ProviderItem};
 
 use anni_provider::cache::{Cache, CachePool};
 use anni_provider::fs::LocalFileSystemProvider;
 use anni_provider::providers::drive::DriveProviderSettings;
 use anni_provider::providers::{CommonConventionProvider, CommonStrictProvider, DriveProvider};
-use anni_provider::{AnniProvider, RepoDatabaseRead};
-use anni_repo::{setup_git2, RepositoryManager};
+use anni_provider::AnniProvider;
+use annil::metadata::MetadataConfig;
+use annil::provider::AnnilProvider;
 use annil::route::admin;
 use annil::route::user;
 use annil::state::{AnnilKeys, AnnilProviders, AnnilState};
@@ -21,82 +20,16 @@ use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
 use tokio::sync::RwLock;
 
-struct LazyDb {
-    metadata: MetadataConfig,
-    db_path: Option<PathBuf>,
-}
-
-impl LazyDb {
-    pub fn new(metadata: &MetadataConfig) -> Self {
-        Self {
-            metadata: metadata.clone(),
-            db_path: None,
-        }
-    }
-
-    pub fn open(&mut self) -> anyhow::Result<RepoDatabaseRead> {
-        let db = match self.db_path {
-            Some(ref p) => p,
-            None => {
-                let p = init_metadata(&self.metadata)?;
-                self.db_path.insert(p)
-            }
-        };
-        Ok(RepoDatabaseRead::new(db)?)
-    }
-}
-
-fn init_metadata(metadata: &MetadataConfig) -> anyhow::Result<PathBuf> {
-    log::info!("Fetching metadata repository...");
-    let repo_root = metadata.base.join("repo");
-    let repo = if !repo_root.exists() {
-        log::debug!("Cloning metadata repository from {}", metadata.repo);
-        RepositoryManager::clone(&metadata.repo, repo_root)?
-    } else if metadata.pull {
-        log::debug!(
-            "Updating metadata repository at branch: {}",
-            metadata.branch
-        );
-        RepositoryManager::pull(repo_root, &metadata.branch)?
-    } else {
-        log::debug!("Loading metadata repository at {}", repo_root.display());
-        RepositoryManager::new(repo_root)?
-    };
-    log::debug!("Generating metadata database...");
-    let repo = repo.into_owned_manager()?;
-    let database_path = metadata.base.join("repo.db");
-    repo.to_database(&database_path)?;
-    log::info!("Metadata repository fetched.");
-
-    Ok(database_path)
-}
-
 async fn init_state(config: Config) -> anyhow::Result<(AnnilState, AnnilProviders, AnnilKeys)> {
-    // proxy settings
-    let mut db = if let Some(metadata) = &config.metadata {
-        if let Some(proxy) = &metadata.proxy {
-            // if metadata.proxy is an empty string, do not use proxy
-            if proxy.is_empty() {
-                setup_git2(None);
-            } else {
-                // otherwise, set proxy in config file
-                setup_git2(Some(proxy.clone()));
-            }
-            // if no proxy was provided, use default behavior (http_proxy)
-        }
-
-        // init metadata
-        Some(LazyDb::new(metadata))
-    } else {
-        None
-    };
+    #[cfg(feature = "metadata")]
+    let mut db = config.metadata.clone().map(MetadataConfig::into_db);
 
     log::info!("Start initializing providers...");
     let now = SystemTime::now();
     let mut providers = Vec::with_capacity(config.providers.len());
     let mut caches = HashMap::new();
 
-    for (provider_name, provider_config) in config.providers.iter().filter(|(_, cfg)| cfg.enable) {
+    for (provider_name, provider_config) in config.providers.iter() {
         log::debug!("Initializing provider: {}", provider_name);
         let mut provider: Box<dyn AnniProvider + Send + Sync> =
             match (&provider_config.item, &mut db) {
@@ -197,18 +130,17 @@ async fn init_state(config: Config) -> anyhow::Result<(AnnilState, AnnilProvider
         if let Some(cache) = provider_config.cache() {
             log::debug!(
                 "Cache configuration detected: root = {}, max-size = {}",
-                cache.root(),
+                cache.root,
                 cache.max_size
             );
-            if !caches.contains_key(cache.root()) {
+            if !caches.contains_key(&cache.root) {
                 // new cache pool
-                let pool = CachePool::new(cache.root(), cache.max_size);
-                caches.insert(cache.root().to_string(), Arc::new(pool));
+                let pool = CachePool::new(&cache.root, cache.max_size);
+                caches.insert(cache.root.to_string(), Arc::new(pool));
             }
-            provider = Box::new(Cache::new(provider, caches[cache.root()].clone()));
+            provider = Box::new(Cache::new(provider, caches[&cache.root].clone()));
         }
-        let provider =
-            AnnilProvider::new(provider_name.to_string(), provider, provider_config.enable).await?;
+        let provider = AnnilProvider::new(provider_name.to_string(), provider).await?;
         providers.push(provider);
     }
     log::info!(
@@ -216,13 +148,13 @@ async fn init_state(config: Config) -> anyhow::Result<(AnnilState, AnnilProvider
         now.elapsed().unwrap()
     );
 
-    // etag
-    let etag = compute_etag(&providers).await;
+    let providers = AnnilProviders(RwLock::new(providers));
+    let etag = providers.compute_etag().await;
 
     // key
-    let sign_key = HS256Key::from_bytes(config.server.key().as_ref());
-    let share_key = HS256Key::from_bytes(config.server.share_key().as_ref())
-        .with_key_id(config.server.share_key_id());
+    let sign_key = HS256Key::from_bytes(config.server.sign_key.as_ref());
+    let share_key = HS256Key::from_bytes(config.server.share_key.as_ref())
+        .with_key_id(&config.server.share_key_id);
     let version = format!("Annil v{}", env!("CARGO_PKG_VERSION"));
     let last_update = SystemTime::now()
         .duration_since(UNIX_EPOCH)
@@ -235,11 +167,11 @@ async fn init_state(config: Config) -> anyhow::Result<(AnnilState, AnnilProvider
             last_update: RwLock::new(last_update),
             etag: RwLock::new(etag),
         },
-        AnnilProviders(RwLock::new(providers)),
+        providers,
         AnnilKeys {
             sign_key,
             share_key,
-            admin_token: config.server.admin_token().to_string(),
+            admin_token: config.server.admin_token,
         },
     ))
 }
@@ -256,7 +188,7 @@ async fn main() -> anyhow::Result<()> {
             .nth(1)
             .unwrap_or_else(|| "config.toml".to_owned()),
     )?;
-    let listen: SocketAddr = config.server.listen().parse()?;
+    let listen: SocketAddr = config.server.listen.parse()?;
     let (state, providers, keys) = init_state(config).await?;
 
     let app = Router::new()
@@ -290,19 +222,97 @@ async fn main() -> anyhow::Result<()> {
     //                 .send_wildcard(),
     //         )
     //         .wrap(Logger::default().exclude("/info"))
-    //         .service(info)
-    //         .service(admin::reload)
-    //         .service(admin::sign)
-    //         .service(cover)
-    //         .service(
-    //             web::resource("/{album_id}/{disc_id}/{track_id}")
-    //                 .route(web::get().to(audio))
-    //                 .route(web::head().to(audio_head)),
-    //         )
-    //         .service(albums)
     // })
     // .bind(listen)?
     // .run()
     // .await?;
     Ok(())
+}
+
+mod config {
+    use annil::metadata::MetadataConfig;
+    use serde::Deserialize;
+    use std::collections::HashMap;
+    use std::fs;
+    use std::path::{Path, PathBuf};
+
+    #[derive(Deserialize)]
+    pub struct Config {
+        pub server: ServerConfig,
+        pub metadata: Option<MetadataConfig>,
+        #[serde(rename = "backends")]
+        pub providers: HashMap<String, ProviderConfig>,
+    }
+
+    impl Config {
+        pub fn from_file<P: AsRef<Path>>(config_path: P) -> anyhow::Result<Self> {
+            let string = fs::read_to_string(config_path)?;
+            let result = toml_edit::easy::from_str(&string)?;
+            Ok(result)
+        }
+    }
+
+    #[derive(Deserialize)]
+    #[serde(rename_all = "kebab-case")]
+    pub struct ServerConfig {
+        /// Server name
+        pub name: String,
+        /// Port to listen on
+        pub listen: String,
+        /// HMAC key for JWT
+        #[serde(rename = "hmac-key")]
+        pub sign_key: String,
+        pub share_key: String,
+        pub share_key_id: String,
+        /// Password to reload data
+        pub admin_token: String,
+    }
+
+    #[derive(Deserialize)]
+    pub struct ProviderConfig {
+        #[serde(flatten)]
+        pub item: ProviderItem,
+        cache: Option<CacheConfig>,
+    }
+
+    impl ProviderConfig {
+        #[inline]
+        pub fn cache(&self) -> Option<&CacheConfig> {
+            self.cache.as_ref()
+        }
+    }
+
+    #[derive(Deserialize)]
+    #[serde(tag = "type")]
+    pub enum ProviderItem {
+        #[serde(rename = "file")]
+        #[serde(rename_all = "kebab-case")]
+        File {
+            root: String,
+            strict: bool,
+            #[serde(default = "default_layer")]
+            layer: usize,
+        },
+        #[serde(rename = "drive")]
+        #[serde(rename_all = "kebab-case")]
+        Drive {
+            corpora: String,
+            drive_id: Option<String>,
+            initial_token_path: Option<PathBuf>,
+            token_path: PathBuf,
+            #[serde(default)]
+            strict: bool,
+        },
+    }
+
+    const fn default_layer() -> usize {
+        2
+    }
+
+    #[derive(Deserialize)]
+    pub struct CacheConfig {
+        pub root: String,
+        #[serde(default, rename = "max-size")]
+        pub max_size: usize,
+    }
 }
