+++ annil/src/state.rs
@@ -1,4 +1,3 @@
-use crate::config::MetadataConfig;
 use crate::provider::AnnilProvider;
 use jwt_simple::prelude::HS256Key;
 use std::ops::Deref;
@@ -13,6 +12,24 @@ pub struct AnnilKeys {
 
 pub struct AnnilProviders(pub RwLock<Vec<AnnilProvider>>);
 
+impl AnnilProviders {
+    pub async fn compute_etag(&self) -> String {
+        let providers = self.0.read().await;
+
+        let mut etag = 0;
+        for provider in providers.iter() {
+            for album in provider.albums().await {
+                if let Ok(uuid) = uuid::Uuid::parse_str(album.as_ref()) {
+                    etag ^= uuid.as_u128();
+                } else {
+                    log::error!("Failed to parse uuid: {album}");
+                }
+            }
+        }
+        format!(r#""{}""#, base64::encode(etag.to_be_bytes()))
+    }
+}
+
 impl Deref for AnnilProviders {
     type Target = RwLock<Vec<AnnilProvider>>;
 
@@ -26,5 +43,6 @@ pub struct AnnilState {
     pub last_update: RwLock<u64>,
     pub etag: RwLock<String>,
 
-    pub metadata: Option<MetadataConfig>,
+    #[cfg(feature = "metadata")]
+    pub metadata: Option<crate::metadata::MetadataConfig>,
 }
