+++ annil/src/utils.rs
@@ -1,4 +1,3 @@
-use crate::provider::AnnilProvider;
 use axum::response::{IntoResponse, IntoResponseParts};
 
 pub(crate) enum Either<L, R> {
@@ -35,18 +34,4 @@ where
             Either::Right(r) => r.into_response(),
         }
     }
-}
-
-pub async fn compute_etag(providers: &[AnnilProvider]) -> String {
-    let mut etag = 0;
-    for provider in providers {
-        for album in provider.albums().await {
-            if let Ok(uuid) = uuid::Uuid::parse_str(album.as_ref()) {
-                etag ^= uuid.as_u128();
-            } else {
-                log::error!("Failed to parse uuid: {album}");
-            }
-        }
-    }
-    format!(r#""{}""#, base64::encode(etag.to_be_bytes()))
 }
